+++ docs/document.sh
@@ -1,18 +1,18 @@
+#!/bin/bash
 
-
-function compile_doxygen()
+compile_doxygen()
 {
     # remove previous versions
-    rm -rf ../${1}
-    mkdir ../${1}
+    rm -rf output
+    mkdir output
 
     # run doxygen
     doxygen Doxyfile_${1}
 
     # if it contains a LaTeX version, compile as .pdf
-    if [ -d ../${1}/latex ];
+    if [ -d output/latex ];
     then
-        cd ../${1}/latex
+        cd output/latex
 
         # build the LaTeX output
         make
@@ -21,14 +21,12 @@ function compile_doxygen()
         cp -R refman.pdf ../${1}-API.pdf
 
         # remove LaTeX files
-        cd ..
-        rm -rf latex
+        cd ../..
+        rm -rf output/latex
     fi
-
-    # go back to scripts folder
-    cd ../doxygen
 }
 
-
 compile_doxygen "RoutinesRGB"
-compile_doxygen "LightingGUI"
+
+
+